Split wbt to seed up CI job execution.
