+++ eng/testing/scenarios/BuildWasmAppsJobsList.txt
@@ -6,6 +6,7 @@ Wasm.Build.NativeRebuild.Tests.SimpleSourceChangeRebuildTest
 Wasm.Build.Templates.Tests.NativeBuildTests
 Wasm.Build.Tests.Blazor.AppsettingsTests
 Wasm.Build.Tests.Blazor.BuildPublishTests
+Wasm.Build.Tests.Blazor.BuildPublishTestsDefaultTemplate
 Wasm.Build.Tests.Blazor.CleanTests
 Wasm.Build.Tests.Blazor.MiscTests
 Wasm.Build.Tests.Blazor.MiscTests2
@@ -15,6 +16,8 @@ Wasm.Build.Tests.BuildPublishTests
 Wasm.Build.Tests.ConfigSrcTests
 Wasm.Build.Tests.HybridGlobalizationTests
 Wasm.Build.Tests.IcuShardingTests
+Wasm.Build.Tests.IcuShardingTests2
+Wasm.Build.Tests.IcuTests
 Wasm.Build.Tests.InvariantGlobalizationTests
 Wasm.Build.Tests.InvariantTimezoneTests
 Wasm.Build.Tests.MainWithArgsTests
